fix repaints etc.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@5266 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/qt2/QContentPane.C
@@ -160,7 +160,7 @@ void QContentPane::mouseDoubleClickEvent(QMouseEvent * e)
 	dc_event_ = double_click(e);
 
 	// doubleClickInterval() is just too long.
-	QTimer::singleShot(QApplication::doubleClickInterval() / 1.5,
+	QTimer::singleShot(int(QApplication::doubleClickInterval() / 1.5),
 		this, SLOT(doubleClickTimeout()));
 }
